Break out via and rule parsers into separate files
